+++ packages/gallery/src/index.js
@@ -4,5 +4,7 @@ import ProGallery from './components/gallery/proGallery';
 import LoveButton from './components/item/loveButton/loveButton';
 import GalleryItem from './components/item/galleryItem';
 import gallerySettings from './settings';
+import ExpandableProGallery from './components/gallery/withFullscreen';
+import GALLERY_EVENTS from './common/constants';
 
-export { ProGallery, LoveButton, GalleryItem, gallerySettings };
+export { ProGallery, ExpandableProGallery, GALLERY_EVENTS, LoveButton, GalleryItem, gallerySettings };